# Ciyu-Video-Client/ciyu/lib/apis/api_spec.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:44333
    description: Generated server url
paths:
...
  /api/v1/sentences/generated:
    get:
      tags:
        - sentence-controller
      operationId: generateCompletions
      parameters:
        - name: beforeText
          in: query
          required: true
          schema:
            type: string
        - name: excludedWordIds
          in: query
          required: true
          schema:
            type: array
            items:
              type: string
        - name: page
          in: query
          description: Zero-based page index (0..N)
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: The size of the page to be returned
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
          schema:
            type: array
            items:
              type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
...
  /api/v1/articles:
    get:
      tags:
        - article-controller
      operationId: getBestArticles
      parameters:
        - name: excludedArticleIds
          in: query
          required: true
          schema:
            type: array
            items:
              type: string
        - name: source
          in: query
          required: true
          schema:
            type: string
        - name: page
          in: query
          description: Zero-based page index (0..N)
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: The size of the page to be returned
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
          schema:
            type: array
            items:
              type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SearchHitsArticle'
  /api/v1/articles/word:
    get:
      tags:
        - article-controller
      operationId: searchByWordId
      parameters:
        - name: wordId
          in: query
          required: true
          schema:
            type: string
        - name: page
          in: query
          description: Zero-based page index (0..N)
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: The size of the page to be returned
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
          schema:
            type: array
            items:
              type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SearchHitsArticle'
components:
  schemas:
...
    Article:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        title:
          type: string
        thumbnail:
          type: string
...
    AggregationsContainerObject:
      type: object
    EntryOption:
      type: object
      properties:
        text:
          type: string
        offset:
          type: integer
          format: int32
        length:
          type: integer
          format: int32
        options:
          type: array
          items:
            $ref: '#/components/schemas/Option'
    Explanation:
      type: object
      properties:
        match:
          type: boolean
        value:
          type: number
          format: double
        description:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/Explanation'
    NestedMetaData:
      type: object
      properties:
        field:
          type: string
        offset:
          type: integer
          format: int32
        child:
          $ref: '#/components/schemas/NestedMetaData'
    Option:
      type: object
      properties:
        text:
          type: string
        highlighted:
          type: string
        score:
          type: number
          format: float
        collateMatch:
          type: boolean
    SearchHitArticle:
      type: object
      properties:
        index:
          type: string
        id:
          type: string
        score:
          type: number
          format: float
        sortValues:
          type: array
          items:
            type: object
        content:
          $ref: '#/components/schemas/Article'
        highlightFields:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        innerHits:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/SearchHitsObject'
        nestedMetaData:
          $ref: '#/components/schemas/NestedMetaData'
        routing:
          type: string
        explanation:
          $ref: '#/components/schemas/Explanation'
        matchedQueries:
          type: array
          items:
            type: string
    SearchHitObject:
      type: object
      properties:
        index:
          type: string
        id:
          type: string
        score:
          type: number
          format: float
        sortValues:
          type: array
          items:
            type: object
        content:
          type: object
        highlightFields:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        nestedMetaData:
          $ref: '#/components/schemas/NestedMetaData'
        routing:
          type: string
        explanation:
          $ref: '#/components/schemas/Explanation'
        matchedQueries:
          type: array
          items:
            type: string
    SearchHitsArticle:
      type: object
      properties:
        searchHits:
          type: array
          items:
            $ref: '#/components/schemas/SearchHitArticle'
        totalHitsRelation:
          type: string
          enum:
            - EQUAL_TO
            - GREATER_THAN_OR_EQUAL_TO
            - "OFF"
        aggregations:
          $ref: '#/components/schemas/AggregationsContainerObject'
        totalHits:
          type: integer
          format: int64
        suggest:
          $ref: '#/components/schemas/Suggest'
        maxScore:
          type: number
          format: float
        empty:
          type: boolean
    SearchHitsObject:
      type: object
      properties:
        searchHits:
          type: array
          items:
            $ref: '#/components/schemas/SearchHitObject'
        totalHitsRelation:
          type: string
          enum:
            - EQUAL_TO
            - GREATER_THAN_OR_EQUAL_TO
            - "OFF"
        aggregations:
          $ref: '#/components/schemas/AggregationsContainerObject'
        totalHits:
          type: integer
          format: int64
        suggest:
          $ref: '#/components/schemas/Suggest'
        maxScore:
          type: number
          format: float
        empty:
          type: boolean
    Suggest:
      type: object
      properties:
        suggestions:
          type: array
          items:
            $ref: '#/components/schemas/SuggestionEntryOption'
    SuggestionEntryOption:
      type: object
      properties:
        name:
          type: string
        size:
          type: integer
          format: int32
        entries:
          type: array
          items:
            $ref: '#/components/schemas/EntryOption'
...